+++ tests/unit/test_client.py
@@ -290,7 +290,7 @@ def test_vault_client_base__merge_secrets():
     }
 
 
-def test_vault_client_base_get_all():
+def test_vault_client_base_get_all_secrets():
     class TestVaultClient(client.VaultClientBase):
         def __init__(self):
             pass
@@ -301,14 +301,14 @@ def test_vault_client_base_get_all():
         def get_secret(self, path):
             return {"a/c": "secret-ac", "b": "secret-b"}[path]
 
-    result = TestVaultClient().get_all(["a", ""])
+    result = TestVaultClient().get_all_secrets(["a", ""])
 
     assert result == {"a": {"c": "secret-ac"}, "b": "secret-b"}
 
-    result = TestVaultClient().get_all(["a"])
+    result = TestVaultClient().get_all_secrets(["a"])
 
     assert result == {"a": {"c": "secret-ac"}}
 
-    result = TestVaultClient().get_all(["a", ""], merged=True)
+    result = TestVaultClient().get_all_secrets(["a", ""], merged=True)
 
     assert result == {"c": "secret-ac", "b": "secret-b"}